test: Extract a reusable conference test harness.
Move ColibriAndJingleXmppConnection and RemoteParticipant out of
ConferenceTest into their own file, and extract the JitsiMeetConferenceImpl
wiring (mock JicofoServices, XMPP provider, chat room, jingle session
tracking) into a ConferenceHarness that new conference-level tests can
reuse instead of rebuilding the mock tower by hand.
